Add files via upload
adding final project report

"Our team has produced five (5) different player (or agent) algorithms for the game, Mancala. We were able to conduct testing and trial simulation for all five algorithms and our measured outcome is shown below under Results. Mancala is a competitive 2 player game that is deterministic. The state space for the game can be upper bounded by about 12 (12 dishes, 48 stones), so the game cannot be 48 determined by searching the entire tree. So, the goal of this project is to compare different AI methods and compete the methods against each other"